+++ addon/components/object-list-view-row.js
@@ -311,12 +311,7 @@ export default FlexberryBaseComponent.extend({
     onRowClick(record, params, e) {
       if (!Ember.isBlank(e)) {
         Ember.set(params, 'originalEvent', Ember.$.event.fix(e));
-      }
-
-      // If user clicked on hierarchy expand button on lookup form we should not process row clicking.
-      let classOfHierarchyExpandButton = 'hierarchy-expand';
-      if (Ember.isBlank(e) || !Ember.$(Ember.get(params, 'originalEvent.target')).hasClass(classOfHierarchyExpandButton))
-      {
+      } else {
         this.sendAction('rowClick', record, params);
       }
     }