+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [2.4.0-beta.0] - 2020-03-18
 ### Fixed
 * Overridden methods in the `LogService` now return promises, as described in the [documentation](https://flexberry.github.io/en/ef_log-service.html).
+* `flexberry-objectlistview` component:
+    * Row button add at mobile version.
 
 ### Changed
 * Update dependency on `ember-flexberry-data` to version `2.4.0-beta.0`.